Massive refactor part1
