+++ src/Pirouette/Term/Symbolic/Prover/Runner.hs
@@ -7,6 +7,7 @@ module Pirouette.Term.Symbolic.Prover.Runner where
 import Control.Monad.Reader
 import Pirouette.Monad
 import Pirouette.SMT
+import Pirouette.SMT.Constraints
 import Pirouette.Term.Symbolic.Eval
 import Pirouette.Term.Symbolic.Prover
 import Pirouette.Term.Syntax.Base
